Relocate netlink information gathering
