+++ backends/fireflyb/cache.go
@@ -3,7 +3,7 @@ package fireflyb
 import (
 	"sync"
 
-	"github.com/scitags/flowd-go/netlink"
+	"github.com/scitags/flowd-go/enrichment/netlink"
 )
 
 type connectionCache struct {
