+++ enrichment/netlink/types.go
