+++ enrichment/netlink/utils.go
