+++ types/types.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/scitags/flowd-go/netlink"
+	"github.com/scitags/flowd-go/enrichment/netlink"
 	"golang.org/x/sys/unix"
 )
